mount: move FindFileSystem to pkg/mount
