+++ cmds/exp/cpu/cpu.go
@@ -24,8 +24,8 @@ import (
 	// It can not, however, unpack password-protected keys yet.
 	"github.com/gliderlabs/ssh"
 	"github.com/kr/pty" // TODO: get rid of krpty
+	"github.com/u-root/u-root/pkg/mount"
 	"github.com/u-root/u-root/pkg/termios"
-	"github.com/u-root/u-root/pkg/uroot/util"
 	// We use this ssh because it can unpack password-protected private keys.
 	ossh "golang.org/x/crypto/ssh"
 	"golang.org/x/sys/unix"
@@ -180,7 +180,7 @@ func runRemote(cmd, port9p string) error {
 	}
 
 	var overlaid bool
-	if util.FindFileSystem("overlay") == nil {
+	if mount.FindFileSystem("overlay") == nil {
 		if err := unix.Mount("overlay", "/tmp/root", "overlay", unix.MS_MGC_VAL, "lowerdir=/tmp/cpu,upperdir=/tmp/local,workdir=/tmp/merge"); err == nil {
 			//overlaid = true
 		} else {
