+++ pkg/mount/fs_linux_root.go
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-package util
+package mount
 
 import (
 	"fmt"
@@ -10,15 +10,13 @@ import (
 )
 
 func TestFindFileSystem(t *testing.T) {
-	var tests = []struct {
+	for _, tt := range []struct {
 		name string
 		err  string
 	}{
 		{"rootfs", "<nil>"},
 		{"bogusfs", "bogusfs not found"},
-	}
-
-	for _, tt := range tests {
+	} {
 		t.Run(tt.name, func(t *testing.T) {
 			err := FindFileSystem(tt.name)
 			// There has to be a better way to do this.
